requirements.txt
accelerate==1.8.1
annotated-types==0.7.0
audioread==3.0.1
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
chatterbox-tts==0.1.2
click==8.2.1
conformer==0.3.2
decorator==4.4.2
diffusers==0.29.0
distlib==0.3.9
einops==0.8.1
filelock==3.18.0
fsspec==2025.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
hf-xet==1.1.5
httplib2==0.22.0
huggingface-hub==0.33.1
identify==2.6.12
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib-metadata==8.7.0
jinja2==3.1.6
joblib==1.5.1
lazy-loader==0.4
librosa==0.11.0
llvmlite==0.44.0
markupsafe==3.0.2
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.1
networkx==3.5
nltk==3.9.1
nodeenv==1.9.1
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnx==1.18.0
packaging==25.0
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
pre-commit==4.2.0
proglog==0.1.12
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydub==0.25.1
pyparsing==3.2.3
pyside6==6.9.1
pyside6-addons==6.9.1
pyside6-essentials==6.9.1
python-dotenv==1.1.1
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.4
resemble-perth==1.0.1
rsa==4.9.1
s3tokenizer==0.1.7
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
setuptools==80.9.0
shiboken6==6.9.1
soundfile==0.13.1
soxr==0.5.0.post1
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.20.3
torch==2.6.0
torchaudio==2.6.0
tqdm==4.67.1
transformers==4.46.3
triton==3.2.0
typing-extensions==4.14.0
typing-inspection==0.4.1
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.31.2
zipp==3.23.0
